Added support for negative split maps old format

+++ GEMTools/src/gt_input_map_parser.c
@@ -852,10 +852,10 @@ GT_INLINE gt_status gt_imp_parse_split_map_v0(const char** const text_line,gt_ma
   // Link both donor & acceptor
   gt_map_set_base_length(acceptor_map,read_base_length-gt_map_get_base_length(donor_map));
   if (gt_map_get_strand(donor_map)==FORWARD) {
-    gt_map_set_next_block(donor_map,acceptor_map,SPLICE,gt_map_get_position(acceptor_map)-gt_map_get_position(donor_map)+gt_map_get_length(donor_map));
+    gt_map_set_next_block(donor_map,acceptor_map,SPLICE,(int64_t)gt_map_get_position(acceptor_map)-(int64_t)gt_map_get_position(donor_map)+(int64_t)gt_map_get_length(donor_map));
   } else {
     const uint64_t acceptor_map_length = (read_base_length!=UINT64_MAX) ? gt_map_get_length(acceptor_map) : 0;
-    gt_map_set_next_block(acceptor_map,donor_map,SPLICE,gt_map_get_position(donor_map)-gt_map_get_position(acceptor_map)+acceptor_map_length);
+    gt_map_set_next_block(acceptor_map,donor_map,SPLICE,(int64_t)gt_map_get_position(donor_map)-(int64_t)gt_map_get_position(acceptor_map)+(int64_t)acceptor_map_length);
   }
   // Detect multiple donor position
   if (gt_expect_false((**text_line)==GT_MAP_SPLITMAP_OPEN_GEMv0)) { // [30;34]=chr10:F74776624~chr10:F[74790025;74790029]
